Remove old templating code
