+++ rewrite-javascript/rewrite/test/javascript/templating/basic.test.ts
@@ -1,11 +1,11 @@
 /*
  * Copyright 2025 the original author or authors.
  * <p>
- * Licensed under the Moderne Source Available License (the "License");
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  * <p>
- * https://docs.moderne.io/licensing/moderne-source-available-license
+ * https://www.apache.org/licenses/LICENSE-2.0
  * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {pattern, capture, Pattern, Matcher} from "../../../src/javascript/templating2";
+import {pattern, capture, Pattern, Matcher} from "../../../src/javascript";
 
-describe('templating2 basics', () => {
+describe('templating basics', () => {
     describe('capture', () => {
         test('creates a capture with name', () => {
             const c = capture('test');
