+++ rewrite-javascript/rewrite/test/javascript/templating/match.test.ts
@@ -1,11 +1,11 @@
 /*
  * Copyright 2025 the original author or authors.
  * <p>
- * Licensed under the Moderne Source Available License (the "License");
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  * <p>
- * https://docs.moderne.io/licensing/moderne-source-available-license
+ * https://www.apache.org/licenses/LICENSE-2.0
  * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,11 +14,17 @@
  * limitations under the License.
  */
 import {fromVisitor, RecipeSpec} from "../../../src/test";
-import {JavaScriptVisitor, typescript} from "../../../src/javascript";
+import {
+    capture,
+    JavaCoordinates,
+    JavaScriptVisitor,
+    pattern,
+    rewrite,
+    template,
+    typescript
+} from "../../../src/javascript";
 import {J} from "../../../src/java";
-import {capture, pattern, rewrite, template} from "../../../src/javascript/templating2";
 import {createDraft, produce} from "immer";
-import {JavaCoordinates} from "../../../src/javascript/templating";
 import Mode = JavaCoordinates.Mode;
 
 describe('match extraction', () => {
